requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aioch==0.0.2
    # via -r requirements.in
amqp==2.5.2
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.3.2
    # via
    #   -r requirements.in
    #   django
attrs==21.4.0
    # via jsonschema
backoff==1.6.0
    # via posthoganalytics
billiard==3.6.3.0
    # via celery
celery-redbeat==2.0.0
    # via -r requirements.in
celery==4.4.2
    # via
    #   -r requirements.in
    #   celery-redbeat
certifi==2019.11.28
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via cryptography
chardet==3.0.4
    # via requests
clickhouse-driver==0.2.1
    # via
    #   -r requirements.in
    #   aioch
    #   clickhouse-pool
clickhouse-pool==0.5.3
    # via -r requirements.in
cryptography==3.4.7
    # via
    #   kafka-helper
    #   social-auth-core
cssselect==1.1.0
    # via toronado
cssutils==1.0.2
    # via toronado
dataclasses_json==0.5.7
    # via -r requirements.in
defusedxml==0.6.0
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
dj-database-url==0.5.0
    # via -r requirements.in
django-axes==5.9.0
    # via -r requirements.in
django-constance==2.8.0
    # via -r requirements.in
django-cors-headers==3.5.0
    # via -r requirements.in
django-deprecate-fields==0.1.1
    # via -r requirements.in
django-extensions==3.1.2
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-ipware==3.0.2
    # via
    #   django-axes
    #   django-structlog
django-loginas==0.3.9
    # via -r requirements.in
django-picklefield==3.0.1
    # via -r requirements.in
django-redis==4.12.1
    # via -r requirements.in
git+https://github.com/zapier/django-rest-hooks.git@v1.6.0
    # via -r requirements.in
django-statsd==2.5.2
    # via -r requirements.in
django-structlog==2.1.3
    # via -r requirements.in
django==3.2.12
    # via
    #   -r requirements.in
    #   django-axes
    #   django-cors-headers
    #   django-deprecate-fields
    #   django-extensions
    #   django-filter
    #   django-picklefield
    #   django-redis
    #   django-rest-hooks
    #   django-structlog
    #   djangorestframework
    #   drf-spectacular
djangorestframework-csv==2.1.0
    # via -r requirements.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   djangorestframework-csv
    #   drf-exceptions-hog
    #   drf-extensions
    #   drf-spectacular
dnspython==2.2.1
    # via -r requirements.in
drf-exceptions-hog==0.2.0
    # via -r requirements.in
drf-extensions==0.7.0
    # via -r requirements.in
drf-spectacular==0.21.1
    # via -r requirements.in
future==0.18.2
    # via lzstring
grpcio==1.33.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==2.8
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==1.6.0
    # via -r requirements.in
git+https://github.com/PostHog/infi.clickhouse_orm@bbc5638f1d2ee19bef388a12b4d16fc138354220
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iso8601==0.1.12
    # via infi-clickhouse-orm
isodate==0.6.1
    # via python3-saml
jsonschema==4.4.0
    # via drf-spectacular
kafka-helper==0.2
    # via -r requirements.in
kafka-python==2.0.2
    # via -r requirements.in
kombu==4.6.8
    # via
    #   -r requirements.in
    #   celery
lxml==4.7.1
    # via
    #   python3-saml
    #   toronado
    #   xmlsec
lzstring==1.0.4
    # via -r requirements.in
mimesis==5.2.1
    # via -r requirements.in
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow==3.15.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
monotonic==1.5
    # via posthoganalytics
mypy-extensions==0.4.3
    # via typing-inspect
numpy==1.21.4
    # via -r requirements.in
oauthlib==3.1.0
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via marshmallow
parso==0.8.1
    # via -r requirements.in
pexpect==4.7.0
    # via -r requirements.in
pickleshare==0.7.5
    # via -r requirements.in
posthoganalytics==1.4.7
    # via -r requirements.in
protobuf==3.13.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
ptyprocess==0.6.0
    # via pexpect
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via social-auth-core
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   posthoganalytics
python-statsd==2.1.0
    # via django-statsd
python3-openid==3.1.0
    # via social-auth-core
python3-saml==1.12.0
    # via -r requirements.in
pytz==2021.1
    # via
    #   -r requirements.in
    #   celery
    #   clickhouse-driver
    #   django
    #   infi-clickhouse-orm
    #   tzlocal
pyyaml==6.0
    # via drf-spectacular
redis==3.4.1
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   django-redis
requests-oauthlib==1.3.0
    # via
    #   -r requirements.in
    #   social-auth-core
requests==2.25.1
    # via
    #   -r requirements.in
    #   django-rest-hooks
    #   infi-clickhouse-orm
    #   posthoganalytics
    #   requests-oauthlib
    #   social-auth-core
semantic_version==2.8.5
    # via -r requirements.in
sentry-sdk==1.3.1
    # via -r requirements.in
six==1.14.0
    # via
    #   djangorestframework-csv
    #   grpcio
    #   isodate
    #   posthoganalytics
    #   protobuf
    #   python-dateutil
    #   tenacity
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.1.0
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.3.0
    # via django
statshog==1.0.6
    # via -r requirements.in
structlog==21.2.0
    # via django-structlog
tenacity==6.1.0
    # via celery-redbeat
toronado==0.1.0
    # via -r requirements.in
typing-extensions==4.1.1
    # via typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
tzlocal==2.1
    # via clickhouse-driver
unicodecsv==0.14.1
    # via djangorestframework-csv
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery
whitenoise==5.2.0
    # via -r requirements.in
xmlsec==1.3.12
    # via python3-saml
zipp==3.1.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
